Refactor the main methods
