+++ main.rb
@@ -1,19 +1,8 @@
 #!/usr/bin/env ruby
 require_relative 'app'
+require_relative 'display'
 
-class Main
-  def display_menu
-    puts "\n"
-    puts 'Please choose an option by entering a number'
-    puts '1 - List all books'
-    puts '2 - List all people'
-    puts '3 - Create a person'
-    puts '4 - Create a book'
-    puts '5 - Create a rental'
-    puts '6 - List all rentals for a given person id'
-    puts '7 - exit'
-  end
-
+class Main < Display
   def main # rubocop:disable Metrics/CyclomaticComplexity
     app = App.new
     user_finished = false
